backend/requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.6.0
bleach==3.1.5
boto==2.49.0
boto3==1.14.63
botocore==1.17.63
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
cryptography==3.1
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
docutils==0.15.2
entrypoints==0.3
gensim==3.8.3
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
JPype1==1.0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kiwisolver==1.2.0
konlpy==0.5.2
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.1
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.19.2
oauthlib==3.1.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pdfminer.six==20200726
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7

pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
scipy==1.5.2
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
smart-open==2.1.1
sortedcontainers==2.2.2
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
tweepy==3.9.0
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
webdriver-manager==3.2.2
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.8.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
